+++ src/main/java/net/fabricmc/fabric/api/tag/TagRegistry.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.tags;
+package net.fabricmc.fabric.api.tag;
 
 import net.minecraft.block.Block;
 import net.minecraft.item.Item;
